Move classes to package org.robolectric.shadows.httpclient.
