+++ robolectric-shadows/shadows-httpclient/src/main/java/org/robolectric/shadows/httpclient/HttpResponseStub.java
@@ -1,4 +1,4 @@
-package org.robolectric.tester.org.apache.http;
+package org.robolectric.shadows.httpclient;
 
 import org.apache.http.Header;
 import org.apache.http.HeaderIterator;
