+++ robolectric-shadows/shadows-httpclient/src/test/java/org/robolectric/shadows/httpclient/ParamsParserTest.java
@@ -1,4 +1,4 @@
-package org.robolectric.tester.org.apache.http.impl;
+package org.robolectric.shadows.httpclient;
 
 import org.apache.http.client.methods.HttpDelete;
 import org.apache.http.client.methods.HttpGet;
@@ -6,7 +6,7 @@ import org.apache.http.client.methods.HttpPost;
 import org.apache.http.client.methods.HttpPut;
 import org.apache.http.entity.StringEntity;
 import org.junit.Test;
-import org.robolectric.tester.org.apache.http.ParamsParser;
+import org.robolectric.shadows.httpclient.ParamsParser;
 
 import java.util.Map;
